+++ lib/plugin_api.dart
@@ -1,9 +1,9 @@
 library flutter_map.plugin_api;
 
 export 'package:flutter_map/flutter_map.dart';
-export 'package:flutter_map/src/map/state.dart';
+// ignore: invalid_export_of_internal_element
+export 'package:flutter_map/src/map/controller.dart' show MapControllerImpl;
+export 'package:flutter_map/src/map/flutter_map_state.dart';
 export 'package:flutter_map/src/misc/private/bounds.dart';
 export 'package:flutter_map/src/misc/private/positioned_tap_detector_2.dart';
 export 'package:flutter_map/src/misc/private/util.dart';
-// ignore: invalid_export_of_internal_element
-export 'package:flutter_map/src/map/controller.dart' show MapControllerImpl;
