+++ lib/src/gestures/map_events.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_map/src/misc/point.dart';
+import 'package:flutter_map/src/map/flutter_map_state.dart';
 import 'package:latlong2/latlong.dart';
 
 /// Event sources which are used to identify different types of
@@ -32,68 +32,52 @@ abstract class MapEvent {
   /// Who / what issued the event.
   final MapEventSource source;
 
-  /// Geographical coordinates related to current event.
-  final LatLng center;
-
-  /// Zoom value related to current event.
-  final double zoom;
+  /// The state of the map after the event.
+  final FlutterMapState mapState;
 
   const MapEvent({
     required this.source,
-    required this.center,
-    required this.zoom,
+    required this.mapState,
   });
 }
 
 /// Base event class which is emitted by MapController instance and
 /// includes information about camera movement
+/// TODO: Change name to reflect that this is a base class for map events
+/// which are not partial (e.g start rotate, rotate, end rotate).
 abstract class MapEventWithMove extends MapEvent {
-  /// Target coordinates of point where map is being pointed to
-  final LatLng targetCenter;
-
-  /// Zoom value of point where map is being pointed to
-  final double targetZoom;
+  final FlutterMapState oldMapState;
 
   const MapEventWithMove({
-    required this.targetCenter,
-    required this.targetZoom,
     required super.source,
-    required super.center,
-    required super.zoom,
+    required this.oldMapState,
+    required super.mapState,
   });
 
   /// Returns a subclass of [MapEventWithMove] if the [source] belongs to a
   /// movement event, otherwise returns null.
   static MapEventWithMove? fromSource({
-    required LatLng targetCenter,
-    required double targetZoom,
-    required LatLng oldCenter,
-    required double oldZoom,
+    required FlutterMapState oldMapState,
+    required FlutterMapState mapState,
     required bool hasGesture,
     required MapEventSource source,
     String? id,
   }) =>
       switch (source) {
         MapEventSource.flingAnimationController => MapEventFlingAnimation(
-            center: oldCenter,
-            zoom: oldZoom,
-            targetCenter: targetCenter,
-            targetZoom: targetZoom,
+            oldMapState: oldMapState,
+            mapState: mapState,
             source: source,
           ),
         MapEventSource.doubleTapZoomAnimationController =>
           MapEventDoubleTapZoom(
-            center: oldCenter,
-            zoom: oldZoom,
-            targetCenter: targetCenter,
-            targetZoom: targetZoom,
+            oldMapState: oldMapState,
+            mapState: mapState,
             source: source,
           ),
         MapEventSource.scrollWheel => MapEventScrollWheelZoom(
-            center: oldCenter,
-            zoom: oldZoom,
-            targetCenter: targetCenter,
-            targetZoom: targetZoom,
+            oldMapState: oldMapState,
+            mapState: mapState,
             source: source,
           ),
         MapEventSource.onDrag ||
@@ -102,10 +86,8 @@ abstract class MapEventWithMove extends MapEvent {
         MapEventSource.custom =>
           MapEventMove(
             id: id,
-            center: oldCenter,
-            zoom: oldZoom,
-            targetCenter: targetCenter,
-            targetZoom: targetZoom,
+            oldMapState: oldMapState,
+            mapState: mapState,
             source: source,
           ),
         _ => null,
@@ -120,8 +102,7 @@ class MapEventTap extends MapEvent {
   const MapEventTap({
     required this.tapPosition,
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -132,8 +113,7 @@ class MapEventSecondaryTap extends MapEvent {
   const MapEventSecondaryTap({
     required this.tapPosition,
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -145,8 +125,7 @@ class MapEventLongPress extends MapEvent {
   const MapEventLongPress({
     required this.tapPosition,
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -157,23 +136,17 @@ class MapEventMove extends MapEventWithMove {
 
   const MapEventMove({
     this.id,
-    required LatLng targetCenter,
-    required double targetZoom,
     required super.source,
-    required super.center,
-    required super.zoom,
-  }) : super(
-          targetCenter: targetCenter,
-          targetZoom: targetZoom,
-        );
+    required super.oldMapState,
+    required super.mapState,
+  });
 }
 
 /// Event which is fired when dragging is started
 class MapEventMoveStart extends MapEvent {
   const MapEventMoveStart({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -181,23 +154,17 @@ class MapEventMoveStart extends MapEvent {
 class MapEventMoveEnd extends MapEvent {
   const MapEventMoveEnd({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
 /// Event which is fired when animation started by fling gesture is in progress
 class MapEventFlingAnimation extends MapEventWithMove {
   const MapEventFlingAnimation({
-    required LatLng targetCenter,
-    required double targetZoom,
     required super.source,
-    required super.center,
-    required super.zoom,
-  }) : super(
-          targetCenter: targetCenter,
-          targetZoom: targetZoom,
-        );
+    required super.oldMapState,
+    required super.mapState,
+  });
 }
 
 /// Emits when InteractiveFlags contains fling and there wasn't enough velocity
@@ -205,8 +172,7 @@ class MapEventFlingAnimation extends MapEventWithMove {
 class MapEventFlingAnimationNotStarted extends MapEvent {
   const MapEventFlingAnimationNotStarted({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -214,8 +180,7 @@ class MapEventFlingAnimationNotStarted extends MapEvent {
 class MapEventFlingAnimationStart extends MapEvent {
   const MapEventFlingAnimationStart({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -223,45 +188,33 @@ class MapEventFlingAnimationStart extends MapEvent {
 class MapEventFlingAnimationEnd extends MapEvent {
   const MapEventFlingAnimationEnd({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
 /// Event which is fired when map is double tapped
 class MapEventDoubleTapZoom extends MapEventWithMove {
   const MapEventDoubleTapZoom({
-    required LatLng targetCenter,
-    required double targetZoom,
     required super.source,
-    required super.center,
-    required super.zoom,
-  }) : super(
-          targetCenter: targetCenter,
-          targetZoom: targetZoom,
-        );
+    required super.oldMapState,
+    required super.mapState,
+  });
 }
 
 /// Event which is fired when scroll wheel is used to zoom
 class MapEventScrollWheelZoom extends MapEventWithMove {
   const MapEventScrollWheelZoom({
-    required LatLng targetCenter,
-    required double targetZoom,
     required super.source,
-    required super.center,
-    required super.zoom,
-  }) : super(
-          targetCenter: targetCenter,
-          targetZoom: targetZoom,
-        );
+    required super.oldMapState,
+    required super.mapState,
+  });
 }
 
 /// Event which is fired when animation for double tap gesture is started
 class MapEventDoubleTapZoomStart extends MapEvent {
   const MapEventDoubleTapZoomStart({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
@@ -269,29 +222,20 @@ class MapEventDoubleTapZoomStart extends MapEvent {
 class MapEventDoubleTapZoomEnd extends MapEvent {
   const MapEventDoubleTapZoomEnd({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
 /// Event which is fired when map is being rotated
-class MapEventRotate extends MapEvent {
+class MapEventRotate extends MapEventWithMove {
   /// Custom ID to identify related object(s)
   final String? id;
 
-  /// Current rotation in radians
-  final double currentRotation;
-
-  /// Target rotation in radians
-  final double targetRotation;
-
   const MapEventRotate({
     required this.id,
-    required this.currentRotation,
-    required this.targetRotation,
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.oldMapState,
+    required super.mapState,
   });
 }
 
@@ -299,25 +243,21 @@ class MapEventRotate extends MapEvent {
 class MapEventRotateStart extends MapEvent {
   const MapEventRotateStart({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
 class MapEventRotateEnd extends MapEvent {
   const MapEventRotateEnd({
     required super.source,
-    required super.center,
-    required super.zoom,
+    required super.mapState,
   });
 }
 
-class MapEventNonRotatedSizeChange extends MapEvent {
+class MapEventNonRotatedSizeChange extends MapEventWithMove {
   const MapEventNonRotatedSizeChange({
     required super.source,
-    required CustomPoint<double> previousNonRotatedSize,
-    required CustomPoint<double> nonRotatedSize,
-    required super.center,
-    required super.zoom,
+    required super.oldMapState,
+    required super.mapState,
   });
 }
